refactor: Clean up remaining ACP provider duplication and naming
Address code review findings from /simplify:

Backend:
- Extract handleACPOutput and doSendACPPrompt to acp_common.go,
  eliminating 3-way duplication across output handler files
- Remove unused workingDir field from GeminiCLIAgent and OpenCodeAgent
- Fix inconsistent AvailableModels() locking (add mu to Gemini/OpenCode)
- Fix buildGeminiCLIModels double iteration (single pass)
- Fix buildOpenCodePrimaryAgents unnecessary slice copy (mutate in place)
- Merge identical Gemini+Copilot cases in control response handler
- Rename codex_control.go to control_response.go and rename
  codexControlResponseDisplayText to controlResponseDisplayText
  since it handles all providers, not just Codex
- Fix stale comments referencing specific provider lists

Frontend:
- Add provider-neutral ACPControlContent/ACPControlActions aliases;
  Copilot and Gemini no longer import Gemini-named components
- Remove dead changePermissionMode from OpenCode plugin
