+++ backend/internal/worker/agent/copilot_output_test.go
@@ -19,7 +19,7 @@ func TestHandleCopilotOutput_AgentMessageChunk(t *testing.T) {
 	agent := newCopilotAgentWithSink(sink)
 
 	input := `{"jsonrpc":"2.0","method":"session/update","params":{"sessionId":"s1","update":{"sessionUpdate":"agent_message_chunk","content":{"type":"text","text":"Hello Copilot"}}}}`
-	handleCopilotCLIOutput(agent, parseLine([]byte(input)))
+	agent.HandleOutput([]byte(input))
 
 	if sink.StreamChunkCount() != 1 {
 		t.Fatalf("expected 1 stream chunk, got %d", sink.StreamChunkCount())
@@ -38,7 +38,7 @@ func TestHandleCopilotOutput_RequestPermission(t *testing.T) {
 	agent := newCopilotAgentWithSink(sink)
 
 	input := `{"jsonrpc":"2.0","id":7,"method":"session/request_permission","params":{"sessionId":"s1","options":[{"optionId":"proceed_once","name":"Allow","kind":"allow_once"}],"toolCall":{"toolCallId":"tc-1","title":"shell","kind":"execute"}}}`
-	handleCopilotCLIOutput(agent, parseLine([]byte(input)))
+	agent.HandleOutput([]byte(input))
 
 	if sink.PersistedControlCount() != 1 {
 		t.Fatalf("expected 1 persisted control request, got %d", sink.PersistedControlCount())
@@ -53,7 +53,7 @@ func TestHandleCopilotOutput_ConfigOptionUpdateBroadcastsPermissionMode(t *testi
 	agent := newCopilotAgentWithSink(sink)
 
 	input := `{"jsonrpc":"2.0","method":"session/update","params":{"sessionId":"s1","update":{"sessionUpdate":"config_option_update","configOptions":[{"id":"mode","currentValue":"https://agentclientprotocol.com/protocol/session-modes#plan","options":[{"value":"https://agentclientprotocol.com/protocol/session-modes#agent","name":"Agent"},{"value":"https://agentclientprotocol.com/protocol/session-modes#plan","name":"Plan"}]},{"id":"model","currentValue":"gpt-5.4-mini","options":[{"value":"gpt-5.4","name":"GPT-5.4"},{"value":"gpt-5.4-mini","name":"GPT-5.4 mini"}]}]}}}`
-	handleCopilotCLIOutput(agent, parseLine([]byte(input)))
+	agent.HandleOutput([]byte(input))
 
 	if agent.permissionMode != CopilotCLIModePlan {
 		t.Fatalf("expected mode plan, got %q", agent.permissionMode)
