+++ backend/internal/worker/agent/gemini_output.go
@@ -3,34 +3,23 @@ package agent
 import (
 	"encoding/json"
 	"log/slog"
-
-	leapmuxv1 "github.com/leapmux/leapmux/generated/proto/leapmux/v1"
 )
 
-// handleGeminiCLIOutput processes a single parsed JSONL message from the Gemini CLI ACP server.
-func handleGeminiCLIOutput(a *GeminiCLIAgent, line *parsedLine) {
+func (a *GeminiCLIAgent) handleOutput(line *parsedLine) {
 	slog.Debug("gemini HandleOutput", "agent_id", a.agentID, "method", line.Method, "len", len(line.Raw))
-
-	switch line.Method {
-	case acpMethodSessionUpdate:
-		a.handleSessionUpdate(line.Params)
-	case acpMethodSessionRequestPermission:
-		a.handleRequestPermission(line.ID, line.Raw)
-	default:
-		if err := a.sink.PersistMessage(leapmuxv1.MessageRole_MESSAGE_ROLE_ASSISTANT, line.Raw, SpanInfo{}); err != nil {
-			slog.Error("gemini persist notification", "agent_id", a.agentID, "method", line.Method, "error", err)
-		}
-	}
+	a.handleACPOutput(line, a.handleExtraSessionUpdate)
 }
 
-func (a *GeminiCLIAgent) handleSessionUpdate(params json.RawMessage) {
-	a.handleACPSessionUpdate(params, func(sessionUpdate string, update json.RawMessage) bool {
-		if sessionUpdate == acpUpdateCurrentModeUpdate {
-			a.handleCurrentModeUpdate(update)
-			return true
-		}
-		return false
-	})
+func (a *GeminiCLIAgent) HandleOutput(content []byte) {
+	a.handleOutput(parseLine(content))
+}
+
+func (a *GeminiCLIAgent) handleExtraSessionUpdate(sessionUpdate string, update json.RawMessage) bool {
+	if sessionUpdate == acpUpdateCurrentModeUpdate {
+		a.handleCurrentModeUpdate(update)
+		return true
+	}
+	return false
 }
 
 func (a *GeminiCLIAgent) handleCurrentModeUpdate(update json.RawMessage) {
@@ -45,4 +34,9 @@ func (a *GeminiCLIAgent) handleCurrentModeUpdate(update json.RawMessage) {
 	a.permissionMode = mode.CurrentModeID
 	a.mu.Unlock()
 	a.sink.UpdatePermissionMode(mode.CurrentModeID)
+}
+
+// cancelSession sends a session/cancel notification.
+func (a *GeminiCLIAgent) cancelSession() error {
+	return a.acpCancelSession()
 }
